Allow skipping anchor checking by matching URI
To avoid needing to turn off anchor checking across the entire
documentation allow skipping based on matching the URI against a regex.

Some sites/pages use JavaScript to perform anchor assignment in a
webpage, which would require rendering the page to determine whether
the anchor exists. Allow fine grain control of whether the anchor is
checked based on pattern matching, until such stage as the retrieved
URLs can be passed through an engine for deeper checking on the HTML
doctree.